Commit before pulling
