+++ frontend/myApplication/login/src/main/java/com/example/myapplication/Buyer/BuyerActivity.java
@@ -24,6 +24,7 @@ import com.example.myapplication.Login;
 import com.example.myapplication.R;
 import com.example.myapplication.Seller.SellerActivity;
 import com.example.myapplication.Shared.DrawerBaseActivity;
+import com.example.myapplication.Shared.DummyActivity;
 import com.example.myapplication.Shared.NetworkManager;
 import com.example.myapplication.Shared.NetworkResponder;
 import com.example.myapplication.Shared.Offer;
@@ -80,19 +81,6 @@ public class BuyerActivity extends DrawerBaseActivity implements AdapterView.OnI
         ArrayAdapter<String> diningHallAdapter = new SimpleSpinAdapter(this, buyerBacker.getDiningHalls());
         diningHallSpinner.setAdapter(diningHallAdapter);
 
-        Offer offer = new Offer();
-        offer.diningHallList = new ArrayList<>();
-        offer.diningHallList.add(true);
-        offer.diningHallList.add(false);
-        offer.diningHallList.add(true);
-        offer.diningHallList.add(false);
-        offer.userId = profile.getID();
-        offer.startTime = LocalDateTime.now();
-        offer.endTime = offer.startTime.plusHours(1);
-        offer.price = 420;
-        networkManager.subscribe("/user/queue/sellerUpdate", new tempResponder());
-        networkManager.send("/swipr/updateOffer", offer.generateQuery());
-        networkManager.subscribe("/user/queue/sellerCancel", new tempResponder());
     }
 
     @Override
@@ -126,12 +114,9 @@ public class BuyerActivity extends DrawerBaseActivity implements AdapterView.OnI
      */
     public void launchInterestsActivity(View view) {
         //launch interest sub activity
-        /*
-        Intent intent = new Intent(this, InterestActivity.class);
+        //Intent intent = new Intent(this, InterestActivity.class);
+        Intent intent = new Intent(this, DummyActivity.class);
         startActivity(intent);
-         */
-
-        networkManager.send("/user/queue/sellerCancel", "");
     }
 
     public void search(View view) {
@@ -163,10 +148,4 @@ public class BuyerActivity extends DrawerBaseActivity implements AdapterView.OnI
             spinner.setSelection(index);
     }
 
-    class tempResponder implements NetworkResponder{
-        @Override
-        public void onMessageReceived(String json) {
-            Log.d("here", json);
-        }
-    }
 }
